requirements.txt
numpy==1.20.3
numba==0.48.0
requests==2.25.1
flask==2.0.1
flask-restful==0.3.9
torch==1.4.0
torchvision==0.5.0
librosa==0.7.0
pyyaml==5.4.1
webrtcvad-wheels
python-socketio
aiohttp
aiohttp_cors
aiortc
websockets
asyncio
boto3==1.16.16
boto==2.49.0
pipenv
python-dotenv